refactor(legacy): separate tower from scaffold class
This is still a temporary solution to slightly make the scaffold class cleaner.
